% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolindex.R
\name{tolindex}
\alias{tolindex}
\title{Calculates feed tolerance index}
\usage{
tolindex(a, d, i, p, n)
}
\arguments{
\item{a}{sum of difference of milk output of treatment percent x1 with respect to control}

\item{d}{class proportion: number of samples in x1 and thereof showing decrease in milk production/total samples studied}

\item{i}{class proportion: number of samples in x1 and thereof showing increase in milk production/total samples studied}

\item{p}{class proportion: number of samples in x1 not showing unhealthy symptoms due to feed replacement/total samples studied}

\item{n}{class proportion: number of samples in x1 showing no change in milk output from control}
}
\description{
the index tries to find out how much feed can be replaced in the diet of bovines and cattles so that milk production is optimal and health of the bovines and cattles are not compromised.
}
\author{
Prakash Pradhan and Shamik Guchhait
}
